requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.7
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary==2.8.5
djangorestframework-bulk==0.2.1
django-filter==2.3.0

